Move GitHub code into its own file
